Creation date is written to the db

+++ app/src/main/java/com/example/txnotes/db/NotesEntity.java
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey;
 @Entity(tableName = "notes")
 public class NotesEntity {
 
-    public NotesEntity(String note_text, Integer note_creation_date) {
+    public NotesEntity(String note_text, Long note_creation_date) {
         this.note_text = note_text;
         this.note_creation_date = note_creation_date;
     }
@@ -22,7 +22,7 @@ public class NotesEntity {
     public String note_text;
 
     @ColumnInfo(name = "note_creation_date")
-    public Integer note_creation_date;
+    public Long note_creation_date;
